requirements.txt
fastapi
langfuse
pydantic_settings
uvicorn
openai
python_decouple
beanie
pandas
pyyaml